Splits a file into multiple files (for the purposes of readability).
